pyproject.toml
[project]
name = "fastapi-contrainer-openai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "openai>=1.86.0",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
